enh: move ComputeEPIMask into nilearn module
