+++ niworkflows/interfaces/masks.py
@@ -4,13 +4,10 @@
 import os
 import numpy as np
 import nibabel as nb
-from nilearn.masking import compute_epi_mask
-import scipy.ndimage as nd
 
 from nipype.interfaces import fsl, ants
 from nipype.interfaces.base import (
     File,
-    BaseInterfaceInputSpec,
     traits,
     isdefined,
     InputMultiPath,
@@ -102,80 +99,6 @@ class BrainExtractionRPT(nrc.SegmentationRC, ants.segmentation.BrainExtraction):
         )
 
         return super(BrainExtractionRPT, self)._post_run_hook(runtime)
-
-
-# TODO: move this interface to nipype.interfaces.nilearn
-class _ComputeEPIMaskInputSpec(nrc._SVGReportCapableInputSpec, BaseInterfaceInputSpec):
-    in_file = File(exists=True, desc="3D or 4D EPI file")
-    dilation = traits.Int(desc="binary dilation on the nilearn output")
-
-
-class _ComputeEPIMaskOutputSpec(reporting.ReportCapableOutputSpec):
-    mask_file = File(exists=True, desc="Binary brain mask")
-
-
-class ComputeEPIMask(nrc.SegmentationRC):
-    input_spec = _ComputeEPIMaskInputSpec
-    output_spec = _ComputeEPIMaskOutputSpec
-
-    def _run_interface(self, runtime):
-        orig_file_nii = nb.load(self.inputs.in_file)
-        in_file_data = orig_file_nii.get_fdata()
-
-        # pad the data to avoid the mask estimation running into edge effects
-        in_file_data_padded = np.pad(
-            in_file_data, (1, 1), "constant", constant_values=(0, 0)
-        )
-
-        padded_nii = nb.Nifti1Image(
-            in_file_data_padded, orig_file_nii.affine, orig_file_nii.header
-        )
-
-        mask_nii = compute_epi_mask(padded_nii, exclude_zeros=True)
-
-        mask_data = np.asanyarray(mask_nii.dataobj).astype(np.uint8)
-        if isdefined(self.inputs.dilation):
-            mask_data = nd.morphology.binary_dilation(mask_data).astype(np.uint8)
-
-        # reverse image padding
-        mask_data = mask_data[1:-1, 1:-1, 1:-1]
-
-        # exclude zero and NaN voxels
-        mask_data[in_file_data == 0] = 0
-        mask_data[np.isnan(in_file_data)] = 0
-
-        better_mask = nb.Nifti1Image(
-            mask_data, orig_file_nii.affine, orig_file_nii.header
-        )
-        better_mask.set_data_dtype(np.uint8)
-        better_mask.to_filename("mask_file.nii.gz")
-
-        self._mask_file = os.path.join(runtime.cwd, "mask_file.nii.gz")
-
-        runtime.returncode = 0
-        return super(ComputeEPIMask, self)._run_interface(runtime)
-
-    def _list_outputs(self):
-        outputs = super(ComputeEPIMask, self)._list_outputs()
-        outputs["mask_file"] = self._mask_file
-        return outputs
-
-    def _post_run_hook(self, runtime):
-        """ generates a report showing slices from each axis of an arbitrary
-        volume of in_file, with the resulting binary brain mask overlaid """
-
-        self._anat_file = self.inputs.in_file
-        self._mask_file = self.aggregate_outputs(runtime=runtime).mask_file
-        self._seg_files = [self._mask_file]
-        self._masked = True
-
-        NIWORKFLOWS_LOG.info(
-            'Generating report for nilearn.compute_epi_mask. file "%s", and mask file "%s"',
-            self._anat_file,
-            self._mask_file,
-        )
-
-        return super(ComputeEPIMask, self)._post_run_hook(runtime)
 
 
 class _ACompCorInputSpecRPT(nrc._SVGReportCapableInputSpec, confounds.CompCorInputSpec):
